+++ wgpu-core/src/indirect_validation/draw.rs
@@ -135,9 +135,10 @@ impl Draw {
                 resource_index: 0,
                 count: 1,
             }],
-            buffers: &[unsafe {
-                // SAFETY: We calculated the binding size to fit within the buffer.
-                hal::BufferBinding::new_unchecked(buffer, 0, binding_size)
+            buffers: &[hal::BufferBinding {
+                buffer,
+                offset: 0,
+                size: Some(binding_size),
             }],
             samplers: &[],
             textures: &[],
@@ -683,9 +684,10 @@ fn create_buffer_and_bind_group(
             resource_index: 0,
             count: 1,
         }],
-        buffers: &[unsafe {
-            // SAFETY: We just created the buffer with this size.
-            hal::BufferBinding::new_unchecked(buffer.as_ref(), 0, BUFFER_SIZE)
+        buffers: &[hal::BufferBinding {
+            buffer: buffer.as_ref(),
+            offset: 0,
+            size: Some(BUFFER_SIZE),
         }],
         samplers: &[],
         textures: &[],